requirements.txt
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
matplotlib==3.7.2
pymysql==1.1.0
sqlalchemy==2.0.20
joblib==1.3.2
python-dotenv==1.0.0 